requirements.txt
python-nmap
colorama